remove unnecessary component "AppSettingsPage" fetch fetchAppSettingsData data
